Fix for dumb problem in MaPLCompiler where I had been casting up to uint64 instead of specifying a literal unsigned long.

+++ Compiler/MaPLCompiler/MaPLCompilerHelpers.cpp
@@ -652,27 +652,27 @@ bool typeNameMatchesPrimitiveType(const std::string &typeName) {
 }
 
 void confirmSignedValueFitsInSignedBits(int64_t value, int8_t bits, MaPLFile *file, antlr4::Token *token) {
-    if (value >= (int64_t)1 << (bits-1) ||
-        value < -((int64_t)1 << (bits-1))) {
+    if (value >= 1UL << (bits-1) ||
+        value < -(1UL << (bits-1))) {
         file->logError(token, "A value of "+std::to_string(value)+" is outside the range of numbers that can be represented by a "+std::to_string(bits)+"-bit signed integer.");
     }
 }
 
 void confirmUnsignedValueFitsInSignedBits(u_int64_t value, int8_t bits, MaPLFile *file, antlr4::Token *token) {
-    if (value >= (int64_t)1 << (bits-1)) {
+    if (value >= 1UL << (bits-1)) {
         file->logError(token, "A value of "+std::to_string(value)+" is outside the range of numbers that can be represented by a "+std::to_string(bits)+"-bit signed integer.");
     }
 }
 
 void confirmSignedValueFitsInUnsignedBits(int64_t value, int8_t bits, MaPLFile *file, antlr4::Token *token) {
-    if (value >= (u_int64_t)1 << bits ||
+    if (value >= 1UL << bits ||
         value < 0) {
         file->logError(token, "A value of "+std::to_string(value)+" is outside the range of numbers that can be represented by a "+std::to_string(bits)+"-bit unsigned integer.");
     }
 }
 
 void confirmUnsignedValueFitsInUnsignedBits(u_int64_t value, int8_t bits, MaPLFile *file, antlr4::Token *token) {
-    if (value >= (u_int64_t)1 << bits) {
+    if (value >= 1UL << bits) {
         file->logError(token, "A value of "+std::to_string(value)+" is outside the range of numbers that can be represented by a "+std::to_string(bits)+"-bit unsigned integer.");
     }
 }
